requirements.txt
streamlit
moviepy
pydub
openai
google-cloud-speech
google-cloud-texttospeech
google-cloud-storage
python-dotenv
requests